+++ databricks/sdk/service/sql.py
@@ -6155,12 +6155,10 @@ class SpotInstancePolicy(Enum):
 
     The breakdown of how the EndpointSpotInstancePolicy converts to per cloud configurations is:
 
-    +-------+--------------------------------------+--------------------------------+ | Cloud |
-    COST_OPTIMIZED | RELIABILITY_OPTIMIZED |
-    +-------+--------------------------------------+--------------------------------+ | AWS | On
-    Demand Driver with Spot Executors | On Demand Driver and Executors | | AZURE | On Demand Driver
-    and Executors | On Demand Driver and Executors |
-    +-------+--------------------------------------+--------------------------------+"""
+    - AWS, COST_OPTIMIZED: On Demand Driver with Spot Executors.
+    - AWS, RELIABILITY_OPTIMIZED: On Demand Driver and Executors.
+    - AZURE, COST_OPTIMIZED: On Demand Driver and Executors.
+    - AZURE, RELIABILITY_OPTIMIZED: On Demand Driver and Executors."""
 
     COST_OPTIMIZED = "COST_OPTIMIZED"
     POLICY_UNSPECIFIED = "POLICY_UNSPECIFIED"