+++ tests/test_dbutils.py
@@ -286,10 +286,8 @@ def test_jobs_task_values_get_throws(dbutils):
         assert str(e) == "Must pass debugValue when calling get outside of a job context. debugValue cannot be None."
 
 
-def test_dbutils_proxy_overrides(dbutils, mocker, restorable_env):
-    import os
-
-    os.environ["DATABRICKS_SOURCE_FILE"] = "test_source_file"
+def test_dbutils_proxy_overrides(dbutils, mocker, monkeypatch):
+    monkeypatch.setenv("DATABRICKS_SOURCE_FILE", "test_source_file")
     mocker.patch(
         "databricks.sdk.dbutils.RemoteDbUtils._cluster_id",
         return_value="test_cluster_id",